test_average: update test / plot
Now it plots both the original node-centered discretized test function,
as well as the cell-centered averaging result.

Note: It is quite questionable as to whether one should add this plot to
the repository, as it can be generated by the code. I decided to do it
so that you can see the result directly (even on github), but I'd normally
advice against having such files in the repository.